requirements.txt
litellm>=1.60.8
openai>=1.61.1
python-dotenv
anthropic>=0.45.0
requests
tqdm
pymongo
upstash-redis
upstash-vector
redis
boto3
bson
